models/dt-lightgbm-direct/requirements.txt
enfobench>=0.3.5,<0.4.0
u8darts[notorch]==0.27.0